### 2023-04-24 1071fba
fix for bug in SectorManagement
