+++ examples/foxes/multi_height/run.py
@@ -8,7 +8,6 @@ from foxes.utils.runners import DaskRunner
 
 
 def run_foxes(args):
-
     cks = None if args.nodask else {FV.STATE: args.chunksize}
 
     mbook = foxes.models.ModelBook()
@@ -63,7 +62,6 @@ def run_foxes(args):
 
 
 if __name__ == "__main__":
-
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "-l",
@@ -131,5 +129,4 @@ if __name__ == "__main__":
         n_workers=args.n_workers,
         threads_per_worker=args.threads_per_worker,
     ) as runner:
-
         runner.run(run_foxes, args=(args,))
