+++ tests/foxes/verification/flappy_0_6/row_Jensen_linear_tophat/test_row_Jensen_linear_tophat.py
@@ -11,7 +11,6 @@ thisdir = Path(inspect.getfile(inspect.currentframe())).parent
 
 
 def test():
-
     n_s = 1000
     n_t = 55
     c = 1000
