+++ foxes/input/farm_layout/add_from_csv.py
@@ -77,7 +77,6 @@ def add_from_csv(
     D = turbine_parameters.pop("D", None)
 
     for i in data.index:
-
         s = 1 if turbine_base_name_count_shift else 0
         tname = (
             f"{turbine_base_name}{i+s}" if col_name is None else data.loc[i, col_name]
