+++ foxes/opt/objectives/max_n_turbines.py
@@ -3,6 +3,7 @@ import numpy as np
 from foxes.opt.core.farm_objective import FarmObjective
 import foxes.variables as FV
 
+
 class MaxNTurbines(FarmObjective):
     """
     Maximizes the number of turrbines.
@@ -87,7 +88,9 @@ class MaxNTurbines(FarmObjective):
         if FV.VALID in problem_results and self.check_valid:
             vld = np.sum(problem_results[FV.VALID].to_numpy(), axis=1)
             if np.min(vld) != np.max(vld):
-                raise ValueError(f"Objective '{self.name}': Number of valid turbines is state dependend, counting impossible")
+                raise ValueError(
+                    f"Objective '{self.name}': Number of valid turbines is state dependend, counting impossible"
+                )
             return np.array([vld[0]], dtype=np.float64)
         else:
             return np.array([self.farm.n_turbines], dtype=np.float64)
@@ -118,11 +121,16 @@ class MaxNTurbines(FarmObjective):
         if self.check_valid:
             n_states = problem_results["n_org_states"].to_numpy()
             n_turbines = self.farm.n_turbines
-            vld = problem_results[FV.VALID].to_numpy().reshape(n_pop, n_states, n_turbines)
+            vld = (
+                problem_results[FV.VALID]
+                .to_numpy()
+                .reshape(n_pop, n_states, n_turbines)
+            )
             vld = np.sum(vld, axis=2)
             if np.any(np.min(vld, axis=1) != np.max(vld, axis=1)):
-                raise ValueError(f"Objective '{self.name}': Number of valid turbines is state dependend, counting impossible")
+                raise ValueError(
+                    f"Objective '{self.name}': Number of valid turbines is state dependend, counting impossible"
+                )
             return vld[:, 0, None]
         else:
             return np.full((n_pop, 1), self.farm.n_turbines, dtype=vars_float.dtype)
-

[…]
