### 2023-04-25 d07820c
new turbine model: RotorCentreCalc
