### 2023-04-26 86c3215
dev rotor centre calc
