### 2023-05-04 #5393 @hertzsprung
Sadly I have no capacity to review for JOSS for the foreseeable future.  I'm happy to be removed from the review pool (I couldn't find any documentation about this process).
