### 2023-05-12 #5393 @antviro
Hi @timtroendle,

Topic seems interesting! I can review it, and I may  even be interested in testing the first part for teaching. Nevertheless I could only start in 1.5 weeks. Is it OK?  If the other reviewer candidate could earlier, perhaps is a good idea not to involve me, but otherwise I can.

Best regards,
Antonio
