### 2023-05-12 #5464 @editorialbot
[REVIEW]: FOXES: Farm Optimization and eXtended yield Evaluation Software
**Submitting author:** @SchmJo (<a href="http://orcid.org/0000-0002-8191-8141">Jonas Schmidt</a>)
**Repository:** https://github.com/FraunhoferIWES/foxes
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.0
