### 2023-05-16 FraunhoferIWES/foxes#34 @isdanni
[Paper] Summary & Statement of Purpose
It seems a bit difficult to distinguish the [Summary](https://github.com/FraunhoferIWES/foxes/blob/paper/paper/paper.md#summary) and [SoP](https://github.com/FraunhoferIWES/foxes/blob/paper/paper/paper.md#statement-of-need) sections of the paper. Based on paper requirements:

- **Summary**: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- **A statement of need**: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

The first paragraph of the SoP can be included in the summary section(as well as certain pieces from the [Overview - Foxes](https://github.com/FraunhoferIWES/foxes#overview)) and most of the Summary can be moved to SoP to explain what specific problem this software is trying to solve.

linking https://github.com/openjournals/joss-reviews/issues/5464 for tracking review progress.
