### 2023-05-22 fb3f6d8
Load windIO file in-memory and unpack
