+++ foxes/input/windio/windio.py
@@ -11,8 +11,9 @@ from foxes.models.turbine_types import CpCtFromTwo
 import foxes.constants as FC
 import foxes.variables as FV
 
+from windIO.utils.yml_utils import load_yaml
 
-def read_resource(res_yaml, fixed_vars={}, **kwargs):
+def read_resource(res, fixed_vars={}, **kwargs):
     """
     Reads a WindIO energy resource
 
@@ -32,10 +33,6 @@ def read_resource(res_yaml, fixed_vars={}, **kwargs):
         The uniform states
 
     """
-    res_yaml = Path(res_yaml)
-
-    with open(res_yaml, "r") as file:
-        res = yaml.load(file, Loader=yaml.loader.BaseLoader)
     wres = res["wind_resource"]
 
     wd = np.array(wres["wind_direction"], dtype=FC.DTYPE)
@@ -106,7 +103,7 @@ def read_resource(res_yaml, fixed_vars={}, **kwargs):
     return StatesTable(data, output_vars=ovars, fixed_vars=fixed_vars, **kwargs)
 
 
-def read_site(site_yaml, **kwargs):
+def read_site(site, **kwargs):
     """
     Reads a WindIO site
 
@@ -123,20 +120,12 @@ def read_site(site_yaml, **kwargs):
         The states object
 
     """
-    site_yaml = Path(site_yaml)
-
-    with open(site_yaml, "r") as file:
-        site = yaml.load(file, Loader=yaml.loader.BaseLoader)
-
     res_yaml = site["energy_resource"]
-    if res_yaml[0] == ".":
-        res_yaml = (site_yaml.parent / res_yaml).resolve()
     states = read_resource(res_yaml, **kwargs)
 
     return states
 
-
-def read_farm(farm_yaml, mbook=None, layout=-1, turbine_models=[], **kwargs):
+def read_farm(fdict, mbook=None, layout=-1, turbine_models=[], **kwargs):
     """
     Reads a WindIO wind farm
 
@@ -162,10 +151,6 @@ def read_farm(farm_yaml, mbook=None, layout=-1, turbine_models=[], **kwargs):
 
     """
     mbook = ModelBook() if mbook is None else mbook
-    farm_yaml = Path(farm_yaml)
-
-    with open(farm_yaml, "r") as file:
-        fdict = yaml.load(file, Loader=yaml.loader.BaseLoader)
 
     if isinstance(layout, str):
         layout = fdict["layouts"][layout]
@@ -181,19 +166,8 @@ def read_farm(farm_yaml, mbook=None, layout=-1, turbine_models=[], **kwargs):
     ldata["x"] = x
     ldata["y"] = y
 
-    if "turbines" in fdict:
-        turbines_yaml = fdict["turbines"]
-        if turbines_yaml[0] == ".":
-            turbines_yaml = (farm_yaml.parent / turbines_yaml).resolve()
-
-        with open(turbines_yaml, "r") as file:
-            tdict = yaml.load(file, Loader=yaml.loader.BaseLoader)
-
-        pdict = tdict["performance"]
-
-    else:
-        tdict = fdict
-        pdict = fdict
+    tdict = fdict["turbines"]
+    pdict = tdict["performance"]
 
     ct_ws = np.array(pdict["Ct_curve"]["Ct_wind_speeds"], dtype=FC.DTYPE)
     ct_data = pd.DataFrame(index=range(len(ct_ws)))
@@ -209,11 +183,17 @@ def read_farm(farm_yaml, mbook=None, layout=-1, turbine_models=[], **kwargs):
     H = float(tdict["hub_height"])
 
     mbook.turbine_types["windio_turbine"] = CpCtFromTwo(
[…]
