### 2023-05-30 FraunhoferIWES/foxes#34 @SchmJo
I updated the paper - hope this is what you had in mind? Thanks!
