### 2023-05-30 ebffe4a
introducing k_var option to wakes and kTI
