+++ examples/foxes/single_state/run.py
@@ -22,7 +22,7 @@ if __name__ == "__main__":
     parser.add_argument("--ti", help="The TI value", type=float, default=0.08)
     parser.add_argument("--rho", help="The air density", type=float, default=1.225)
     parser.add_argument(
-        "-dx", "--deltax", help="The turbine distance in x", type=float, default=500.0
+        "-dx", "--deltax", help="The turbine distance in x", type=float, default=800.0
     )
     parser.add_argument(
         "-dy", "--deltay", help="Turbine layout y step", type=float, default=0.0
@@ -34,10 +34,13 @@ if __name__ == "__main__":
         default="NREL-5MW-D126-H90.csv",
     )
     parser.add_argument(
+        "-m", "--tmodels", help="The turbine models", default=["kTI_K1", "kTI_K2"], nargs="+"
+    )
+    parser.add_argument(
         "-w",
         "--wakes",
         help="The wake models",
-        default=["Bastankhah_quadratic", "CrespoHernandez_max"],
+        default=["B_K1", "CH_K2"],
         nargs="+",
     )
     parser.add_argument("-r", "--rotor", help="The rotor model", default="centre")
@@ -45,9 +48,6 @@ if __name__ == "__main__":
         "-p", "--pwakes", help="The partial wakes model", default="rotor_points"
     )
     parser.add_argument("-f", "--frame", help="The wake frame", default="rotor_wd")
-    parser.add_argument(
-        "-m", "--tmodels", help="The turbine models", default=["kTI_02"], nargs="+"
-    )
     parser.add_argument("-v", "--var", help="The plot variable", default=FV.WS)
     parser.add_argument(
         "-it", "--iterative", help="Use iterative algorithm", action="store_true"
@@ -58,6 +58,14 @@ if __name__ == "__main__":
     mbook = foxes.ModelBook()
     ttype = foxes.models.turbine_types.PCtFile(args.turbine_file)
     mbook.turbine_types[ttype.name] = ttype
+    mbook.turbine_models["kTI_K1"] = foxes.models.turbine_models.kTI(kTI=0.2, k_var="K1", ti_var=FV.AMB_TI)
+    mbook.turbine_models["kTI_K2"] = foxes.models.turbine_models.kTI(kTI=0.4, k_var="K2", ti_var=FV.AMB_TI)
+    mbook.wake_models["B_K1"] = foxes.models.wake_models.wind.BastankhahWake(
+        superposition="quadratic", k_var="K1"
+    )
+    mbook.wake_models["CH_K2"] = foxes.models.wake_models.ti.CrespoHernandezTIWake(
+        superposition="max", k_var="K2", use_ambti=False
+    )
 
     # create states
     states = foxes.input.states.SingleStateStates(
