### 2023-05-31 FraunhoferIWES/foxes#34 @isdanni
Thanks! The updated paper fits the requirement well.
