+++ foxes/models/farm_models/__init__.py
@@ -1,1 +1,4 @@
+"""
+Farm models.
+"""
 from .turbine2farm import Turbine2FarmModel
