+++ docs/source/history.md
@@ -273,5 +273,9 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 - Models:
   - Improved: Now option to specify wake growth variable name, such that multiple `kTI` models could be used, resulting in different `k`'s for different wake models
+- Utils:
+  - Improved `DaskRunner`: Now features slurm clusters
+- Examples:
+  - New: `timeseries_slurm`, shows how to run foxes on a HPC with slurm queueing system
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.3.5](https://github.com/FraunhoferIWES/foxes/commits/v0.3.6)
