+++ examples/foxes/timeseries_slurm/README.md
@@ -1,0 +1,4 @@
+# foxes example: _timeseries\_slurm_
+
+Evaluation of a wind farm for a time series of uniform wind and TI states
+on an HPC cluster with SLURM queue system.
