### 2023-06-05 13bf520
model SetFarmVars now supports PopStates
