### 2023-06-07 7733f8d
Resolve "Lookup Table (for ct depending on derating)"
