+++ foxes/input/states/states_table.py
@@ -13,49 +13,30 @@ class StatesTable(States):
     """
     States from a `pandas.DataFrame` or a pandas readable file.
 
-    Parameters
-    ----------
-    data_source : str or pandas.DataFrame
-        Either path to a file or data
-    output_vars : list of str
-        The output variables
-    var2col : dict
-        Mapping from variable names to data column names
-    fixed_vars : dict
-        Fixed uniform variable values, instead of
-        reading from data
-    profiles : dict
-        Key: output variable name str, Value: str or dict
-        or `foxes.core.VerticalProfile`
-    pd_read_pars : dict
-        pandas file reading parameters
-    states_sel : slice or range or list of int, optional
-        States subset selection
-    states_loc : list, optional
-        State index selection via pandas loc function
-
     Attributes
     ----------
-    data_source : str or pandas.DataFrame
+    data_source: str or pandas.DataFrame
         Either path to a file or data
-    ovars : list of str
+    ovars: list of str
         The output variables
-    var2col : dict
+    var2col: dict
         Mapping from variable names to data column names
-    fixed_vars : dict
+    fixed_vars: dict
         Fixed uniform variable values, instead of
         reading from data
-    profdicts : dict
+    profdicts: dict
         Key: output variable name str, Value: str or dict
         or `foxes.core.VerticalProfile`
-    rpars : dict
+    rpars: dict
         pandas file reading parameters
-    states_sel : slice or range or list of int
+    states_sel: slice or range or list of int
         States subset selection
-    states_loc : list
+    states_loc: list
         State index selection via pandas loc function
-    RDICT : dict
+    RDICT: dict
         Default pandas file reading parameters
+    
+    :group: input.states
 
     """
 
@@ -72,6 +53,31 @@ class StatesTable(States):
         states_sel=None,
         states_loc=None,
     ):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        data_source: str or pandas.DataFrame
+            Either path to a file or data
+        output_vars: list of str
+            The output variables
+        var2col: dict
+            Mapping from variable names to data column names
+        fixed_vars: dict
+            Fixed uniform variable values, instead of
+            reading from data
+        profiles: dict
+            Key: output variable name str, Value: str or dict
+            or `foxes.core.VerticalProfile`
+        pd_read_pars: dict
+            pandas file reading parameters
+        states_sel: slice or range or list of int, optional
+            States subset selection
+        states_loc: list, optional
+            State index selection via pandas loc function
+
+        """
         super().__init__()
 
         self.data_source = data_source
@@ -98,11 +104,11 @@ class StatesTable(States):
 
         Parameters
         ----------
-        states_sel : slice or range or list of int, optional
+        states_sel: slice or range or list of int, optional
             States subset selection
-        states_loc : list, optional
+        states_loc: list, optional
             State index selection via pandas loc function
-        verbosity : int
+        verbosity: int
             The verbosity level, 0 = silent
 
         """
@@ -127,14 +133,14 @@ class StatesTable(States):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
-        verbosity : int
+        verbosity: int
             The verbosity level, 0 = silent
 
         Returns
         -------
-        idata : dict
+        idata: dict
             The dict has exactly two entries: `data_vars`,
             a dict with entries `name_str -> (dim_tuple, data_ndarray)`;
             and `coords`, a dict with entries `dim_name_str -> dim_array`
@@ -242,7 +248,7 @@ class StatesTable(States):
 
         Returns
         -------
-        indices : array_like
[…]
