+++ foxes/core/turbine_type.py
@@ -9,36 +9,42 @@ class TurbineType(TurbineModel):
     Rotor diameter and hub height can be overwritten
     by individual settings in the Turbine object.
 
-    Parameters
-    ----------
-    name : str, optional
-        The model name
-    D : float, optional
-        The rotor diameter
-    H : float, optional
-        The hub height
-    P_nominal : float, optional
-        The nominal power in kW
-    P_unit : str
-        The unit of power, choices:
-        W, kW, MW, GW
-
     Attributes
     ----------
-    name : str
+    name: str
         The model name
-    D : float
+    D: float
         The rotor diameter
-    H : float
+    H: float
         The hub height
-    P_nominal : float
+    P_nominal: float
         The nominal power in kW
-    P_unit : str
+    P_unit: str
         The unit of power
+
+    :group: core
 
     """
 
     def __init__(self, name=None, D=None, H=None, P_nominal=None, P_unit="kW"):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        name: str, optional
+            The model name
+        D: float, optional
+            The rotor diameter
+        H: float, optional
+            The hub height
+        P_nominal: float, optional
+            The nominal power in kW
+        P_unit: str
+            The unit of power, choices:
+            W, kW, MW, GW
+
+        """
         super().__init__()
 
         self.name = name if name is not None else type(self).__name__

[…]
