+++ foxes/models/turbine_models/yaw2yawm.py
@@ -9,6 +9,9 @@ class YAW2YAWM(TurbineModel):
     """
     Calculates delta yaw (i.e. YAWM) from absolute
     yaw (i.e. YAW)
+
+    :group: models.turbine_models
+    
     """
 
     def output_farm_vars(self, algo):
@@ -17,12 +20,12 @@ class YAW2YAWM(TurbineModel):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
 
         Returns
         -------
-        output_vars : list of str
+        output_vars: list of str
             The output variable names
 
         """
@@ -37,19 +40,19 @@ class YAW2YAWM(TurbineModel):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
-        mdata : foxes.core.Data
+        mdata: foxes.core.Data
             The model data
-        fdata : foxes.core.Data
+        fdata: foxes.core.Data
             The farm data
-        st_sel : numpy.ndarray of bool
+        st_sel: numpy.ndarray of bool
             The state-turbine selection,
             shape: (n_states, n_turbines)
 
         Returns
         -------
-        results : dict
+        results: dict
             The resulting data, keys: output variable str.
             Values: numpy.ndarray with shape (n_states, n_turbines)
 
