+++ foxes/models/wake_superpositions/product.py
@@ -18,27 +18,33 @@ class ProductSuperposition(WakeSuperposition):
     Source: https://arxiv.org/pdf/2010.03873.pdf
             Equation (8)
 
-    Parameters
+    Attributes
     ----------
-    lim_low : dict, optional
+    lim_low: dict
         Lower limits of the final wake deltas. Key: variable str,
         value: float
-    lim_high : dict, optional
+    lim_high: dict
         Higher limits of the final wake deltas. Key: variable str,
         value: float
 
-    Attributes
-    ----------
-    lim_low : dict
-        Lower limits of the final wake deltas. Key: variable str,
-        value: float
-    lim_high : dict
-        Higher limits of the final wake deltas. Key: variable str,
-        value: float
+    :group: models.wake_superpositions
 
     """
 
     def __init__(self, lim_low=None, lim_high=None):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        lim_low: dict, optional
+            Lower limits of the final wake deltas. Key: variable str,
+            value: float
+        lim_high: dict, optional
+            Higher limits of the final wake deltas. Key: variable str,
+            value: float
+
+        """
         super().__init__()
         self.lim_low = lim_low
         self.lim_high = lim_high
@@ -59,28 +65,28 @@ class ProductSuperposition(WakeSuperposition):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
-        mdata : foxes.core.Data
+        mdata: foxes.core.Data
             The model data
-        fdata : foxes.core.Data
+        fdata: foxes.core.Data
             The farm data
-        states_source_turbine : numpy.ndarray
+        states_source_turbine: numpy.ndarray
             For each state, one turbine index for the
             wake causing turbine. Shape: (n_states,)
-        sel_sp : numpy.ndarray of bool
+        sel_sp: numpy.ndarray of bool
             The selection of points, shape: (n_states, n_points)
-        variable : str
+        variable: str
             The variable name for which the wake deltas applies
-        wake_delta : numpy.ndarray
+        wake_delta: numpy.ndarray
             The original wake deltas, shape: (n_states, n_points)
-        wake_model_result : numpy.ndarray
+        wake_model_result: numpy.ndarray
             The new wake deltas of the selected points,
             shape: (n_sel_sp,)
 
         Returns
         -------
-        wdelta : numpy.ndarray
+        wdelta: numpy.ndarray
             The updated wake deltas, shape: (n_states, n_points)
 
         """
@@ -99,22 +105,22 @@ class ProductSuperposition(WakeSuperposition):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
-        mdata : foxes.core.Data
+        mdata: foxes.core.Data
             The model data
-        fdata : foxes.core.Data
+        fdata: foxes.core.Data
             The farm data
-        variable : str
+        variable: str
             The variable name for which the wake deltas applies
-        amb_results : numpy.ndarray
+        amb_results: numpy.ndarray
             The ambient results, shape: (n_states, n_points)
-        wake_delta : numpy.ndarray
+        wake_delta: numpy.ndarray
             The wake deltas, shape: (n_states, n_points)
 
         Returns
         -------
-        final_wake_delta : numpy.ndarray
+        final_wake_delta: numpy.ndarray
             The final wake delta, which will be added to the ambient
             results by simple plus operation. Shape: (n_states, n_points)
 
