+++ foxes/models/vertical_profiles/__init__.py
@@ -2,5 +2,8 @@
 Verctical profile models.
 """
 from .uniform import UniformProfile
-
-from . import abl_log
+from .abl_log_neutral_ws import ABLLogNeutralWsProfile
+from .abl_log_stable_ws import ABLLogStableWsProfile
+from .abl_log_unstable_ws import ABLLogUnstableWsProfile
+from .abl_log_ws import ABLLogWsProfile
+from .sheared_ws import ShearedProfile
