+++ foxes/models/wake_models/wind/turbopark.py
@@ -11,43 +11,49 @@ class TurbOParkWake(GaussianWakeModel):
 
     https://iopscience.iop.org/article/10.1088/1742-6596/2265/2/022063/pdf
 
-    Parameters
-    ----------
-    superpositions : dict
-        The superpositions. Key: variable name str,
-        value: The wake superposition model name,
-        will be looked up in model book
-    A : float
-        The wake growth parameter A.
-    sbeta_factor : float
-        Factor multiplying sbeta
-    ct_max : float
-        The maximal value for ct, values beyond will be limited
-        to this number
-    c1 : float
-        Factor from Frandsen turbulence model
-    c2 : float
-        Factor from Frandsen turbulence model
-
     Attributes
     ----------
-    A : float
+    A: float
         The wake growth parameter A.
-    sbeta_factor : float
+    sbeta_factor: float
         Factor multiplying sbeta
-    ct_max : float
+    ct_max: float
         The maximal value for ct, values beyond will be limited
         to this number
-    c1 : float
+    c1: float
         Factor from Frandsen turbulence model
-    c2 : float
+    c2: float
         Factor from Frandsen turbulence model
+
+    :group: models.wake_models.wind
 
     """
 
     def __init__(
         self, superposition, A, sbeta_factor=0.25, ct_max=0.9999, c1=1.5, c2=0.8
     ):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        superpositions: dict
+            The superpositions. Key: variable name str,
+            value: The wake superposition model name,
+            will be looked up in model book
+        A: float
+            The wake growth parameter A.
+        sbeta_factor: float
+            Factor multiplying sbeta
+        ct_max: float
+            The maximal value for ct, values beyond will be limited
+            to this number
+        c1: float
+            Factor from Frandsen turbulence model
+        c2: float
+            Factor from Frandsen turbulence model
+
+        """
         super().__init__(superpositions={FV.WS: superposition})
 
         self.A = A
@@ -69,15 +75,15 @@ class TurbOParkWake(GaussianWakeModel):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
-        mdata : foxes.core.Data
+        mdata: foxes.core.Data
             The model data
-        fdata : foxes.core.Data
+        fdata: foxes.core.Data
             The farm data
-        n_points : int
+        n_points: int
             The number of wake evaluation points
-        wake_deltas : dict
+        wake_deltas: dict
             The wake deltas storage, add wake deltas
             on the fly. Keys: Variable name str, for which the
             wake delta applies, values: numpy.ndarray with
@@ -94,24 +100,24 @@ class TurbOParkWake(GaussianWakeModel):
 
         Parameters
         ----------
-        algo : foxes.core.Algorithm
+        algo: foxes.core.Algorithm
             The calculation algorithm
-        mdata : foxes.core.Data
+        mdata: foxes.core.Data
             The model data
-        fdata : foxes.core.Data
+        fdata: foxes.core.Data
             The farm data
-        states_source_turbine : numpy.ndarray
+        states_source_turbine: numpy.ndarray
             For each state, one turbine index for the
             wake causing turbine. Shape: (n_states,)
-        x : numpy.ndarray
+        x: numpy.ndarray
             The x values, shape: (n_states, n_points)
 
         Returns
         -------
-        amsi : tuple
+        amsi: tuple
             The amplitude and sigma, both numpy.ndarray
             with shape (n_sp_sel,)
-        sp_sel : numpy.ndarray of bool
+        sp_sel: numpy.ndarray of bool
             The state-point selection, for which the wake
[…]
