+++ foxes/utils/dict.py
@@ -2,23 +2,29 @@ class Dict(dict):
     """
     A slightly enhanced dictionary.
 
-    Parameters
-    ----------
-    *args : tuple, optional
-        Arguments passed to `dict`
-    name : str, optional
-        The dictionary name
-    **kwargs : dict, optional
-        Arguments passed to `dict`
-
     Attributes
     ----------
-    name : str
+    name: str
         The dictionary name
+    
+    :group: utils
 
     """
 
     def __init__(self, *args, name=None, **kwargs):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        *args: tuple, optional
+            Arguments passed to `dict`
+        name: str, optional
+            The dictionary name
+        **kwargs: dict, optional
+            Arguments passed to `dict`
+
+        """
         super().__init__(*args, **kwargs)
         self.name = name if name is not None else type(self).__name__
 
