+++ foxes/utils/abl/sheared.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 
 def calc_ws(height, height0, WS0, shear):
     """
@@ -7,20 +5,21 @@ def calc_ws(height, height0, WS0, shear):
 
     Parameters
     ----------
-    height : float
+    height: float
         The evaluation height
-    height0 : float
+    height0: float
         Reference height
-    WS0 : float
+    WS0: float
         Reference wind speed
-    shear : float
+    shear: float
         Shear exponent
-
 
     Returns
     -------
-    ws : float
+    ws: float
         The wind speed
+    
+    :group: utils.abl.sheared
 
     """
     return WS0 * (height / height0) ** shear
