+++ foxes/utils/abl/neutral.py
@@ -8,15 +8,17 @@ def logz(height, z0):
 
     Parameters
     ----------
-    height : float
+    height: float
         The evaluation height
-    z0 : float
+    z0: float
         The roughness length
 
     Returns
     -------
-    lz : float
+    lz: float
         The log factor
+    
+    :group: utils.abl.neutral
 
     """
     h = np.maximum(height, z0)
@@ -30,19 +32,21 @@ def ustar(ws_ref, h_ref, z0, kappa=0.41):
 
     Parameters
     ----------
-    ws_ref : float
+    ws_ref: float
         The reference wind speed
-    h_ref : float
+    h_ref: float
         The reference height
-    z0 : float
+    z0: float
         The roughness length
-    kappa : float
+    kappa: float
         The van-Karman constant
 
     Returns
     -------
-    ustar : float
+    ustar: float
         The friction velocity
+    
+    :group: utils.abl.neutral
 
     """
     lz = logz(h_ref, z0)
@@ -55,19 +59,21 @@ def calc_ws(height, z0, ustar, kappa=0.41):
 
     Parameters
     ----------
-    height : float
+    height: float
         The evaluation height
-    z0 : float
+    z0: float
         The roughness length
-    ustar : float
+    ustar: float
         The friction velocity
-    kappa : float
+    kappa: float
         The van-Karman constant
 
     Returns
     -------
-    ws : float
+    ws: float
         The wind speed
+    
+    :group: utils.abl.neutral
 
     """
     return ustar / kappa * logz(height, z0)
