+++ foxes/utils/geom2d/__init__.py
@@ -5,3 +5,4 @@ from .area_geometry import AreaGeometry, InvertedAreaGeometry
 from .area_geometry import AreaUnion, AreaIntersection
 from .polygon import ClosedPolygon
 from .circle import Circle
+from .half_plane import HalfPlane
