+++ docs/source/conf.py
@@ -44,6 +44,7 @@ extensions = [
     "sphinx_immaterial",
     "sphinx_immaterial.apidoc.python.apigen",
     "sphinx.ext.autodoc",
+    'sphinx.ext.autosectionlabel',
     #"sphinx.ext.autosummary",
     #"sphinx.ext.intersphinx",
     #"sphinx.ext.mathjax",
@@ -326,6 +327,7 @@ python_apigen_modules = {
     "foxes.input.farm_layout": "input/farm_layout/",
     "foxes.input.states": "input/states/",
     "foxes.input.windio": "input/windio/",
+    "foxes.models.model_book": "models/model_book/",
     "foxes.models.farm_controllers": "models/farm_controllers/",
     "foxes.models.farm_models": "models/farm_models/",
     "foxes.models.partial_wakes": "models/partial_wakes/",
