+++ foxes/opt/problems/layout/geom_layouts/objectives.py
@@ -34,16 +34,79 @@ class OMaxN(Objective):
         )
 
     def n_components(self):
+        """
+        Returns the number of components of the
+        function.
+
+        Returns
+        -------
+        int:
+            The number of components.
+
+        """
         return 1
 
     def maximize(self):
+        """
+        Returns flag for maximization of each component.
+
+        Returns
+        -------
+        flags: np.array
+            Bool array for component maximization,
+            shape: (n_components,)
+
+        """
         return [True]
 
     def calc_individual(self, vars_int, vars_float, problem_results, cmpnts=None):
+        """
+        Calculate values for a single individual of the
+        underlying problem.
+
+        Parameters
+        ----------
+        vars_int : np.array
+            The integer variable values, shape: (n_vars_int,)
+        vars_float : np.array
+            The float variable values, shape: (n_vars_float,)
+        problem_results : Any
+            The results of the variable application
+            to the problem
+        components : list of int, optional
+            The selected components or None for all
+
+        Returns
+        -------
+        values : np.array
+            The component values, shape: (n_sel_components,)
+
+        """
         __, valid = problem_results
         return np.sum(valid)
 
     def calc_population(self, vars_int, vars_float, problem_results, cmpnts=None):
+        """
+        Calculate values for all individuals of a population.
+
+        Parameters
+        ----------
+        vars_int : np.array
+            The integer variable values, shape: (n_pop, n_vars_int)
+        vars_float : np.array
+            The float variable values, shape: (n_pop, n_vars_float)
+        problem_results : Any
+            The results of the variable application
+            to the problem
+        components : list of int, optional
+            The selected components or None for all
+
+        Returns
+        -------
+        values : np.array
+            The component values, shape: (n_pop, n_sel_components)
+
+        """
         __, valid = problem_results
         return np.sum(valid, axis=1)[:, None]
 
@@ -107,17 +170,80 @@ class OFixN(Objective):
         self.N = N
 
     def n_components(self):
+        """
+        Returns the number of components of the
+        function.
+
+        Returns
+        -------
+        int:
+            The number of components.
+
+        """
         return 1
 
     def maximize(self):
+        """
+        Returns flag for maximization of each component.
+
+        Returns
+        -------
+        flags: np.array
+            Bool array for component maximization,
+            shape: (n_components,)
+
+        """
         return [False]
 
     def calc_individual(self, vars_int, vars_float, problem_results, cmpnts=None):
+        """
+        Calculate values for a single individual of the
+        underlying problem.
+
+        Parameters
+        ----------
+        vars_int : np.array
+            The integer variable values, shape: (n_vars_int,)
+        vars_float : np.array
+            The float variable values, shape: (n_vars_float,)
+        problem_results : Any
+            The results of the variable application
+            to the problem
+        components : list of int, optional
+            The selected components or None for all
+
+        Returns
+        -------
+        values : np.array
+            The component values, shape: (n_sel_components,)
+
+        """
         __, valid = problem_results
         N = np.sum(valid, dtype=np.float64)
         return np.maximum(N - self.N, self.N - N)
 
     def calc_population(self, vars_int, vars_float, problem_results, cmpnts=None):
+        """
+        Calculate values for all individuals of a population.
+
+        Parameters
+        ----------
+        vars_int : np.array
+            The integer variable values, shape: (n_pop, n_vars_int)
+        vars_float : np.array
+            The float variable values, shape: (n_pop, n_vars_float)
+        problem_results : Any
+            The results of the variable application
+            to the problem
+        components : list of int, optional
[…]
