+++ docs/source/api_foxes.rst
@@ -1,0 +1,9 @@
+foxes
+=====
+The top-level package provides direct access to some 
+of selected classes and functions.
+
+All objects listed here can be reached directly
+via the *foxes* package, e.g. *foxes.WindFarm*.
+
+    .. python-apigen-group:: foxes
