+++ docs/source/models/model_book/ModelBook.__init__.rst
@@ -1,0 +1,10 @@
+..
+  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
+
+
+:hide-edit-link:
+
+
+
+.. python-apigen-entity-page::  foxes.models.model_book.ModelBook.__init__
+
