+++ foxes/models/wake_models/induction/__init__.py
@@ -1,0 +1,6 @@
+"""
+Induction wake models
+"""
+
+from .rankine_half_body import RHB
+
