+++ examples/foxes/induction_RHB/run_DEBUG.py
@@ -1,0 +1,179 @@
+import numpy as np
+import argparse
+import matplotlib.pyplot as plt
+
+import foxes
+import foxes.variables as FV
+
+if __name__ == "__main__":
+    # define arguments and options:
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-nt", "--n_t", help="The number of turbines", type=int, default=2
+    )
+    parser.add_argument(
+        "-l",
+        "--layout",
+        help="The wind farm layout file (path or static)",
+        default=None,
+    )
+    parser.add_argument("--ws", help="The wind speed", type=float, default=9.0)
+    parser.add_argument("--wd", help="The wind direction", type=float, default=270.0)
+    parser.add_argument("--ti", help="The TI value", type=float, default=0.08)
+    parser.add_argument("--rho", help="The air density", type=float, default=1.225)
+    parser.add_argument(
+        "-dx", "--deltax", help="The turbine distance in x", type=float, default=800.0
+    )
+    parser.add_argument(
+        "-dy", "--deltay", help="Turbine layout y step", type=float, default=0.
+    )
+    parser.add_argument(
+        "-t",
+        "--turbine_file",
+        help="The P-ct-curve csv file (path or static)",
+        default="NREL-5MW-D126-H90.csv",
+    )
+    parser.add_argument(
+        "-m", "--tmodels", help="The turbine models", default=["kTI_K1", "kTI_K2"], nargs="+"
+    )
+    parser.add_argument(
+        "-w",
+        "--wakes",
+        help="The wake models",
+        default=["RHB_linear"],
+        nargs="+",
+    )
+    parser.add_argument("-r", "--rotor", help="The rotor model", default="centre")
+    parser.add_argument(
+        "-p", "--pwakes", help="The partial wakes model", default="auto"
+    )
+    parser.add_argument("-f", "--frame", help="The wake frame", default="rotor_wd")
+    parser.add_argument("-v", "--var", help="The plot variable", default=FV.WS)
+    parser.add_argument(
+        "-it", "--iterative", help="Use iterative algorithm", action="store_true"
+    )
+    args = parser.parse_args()
+
+    #############################################
+    # OVERWRITE ARGS FOR DEBUGGING HERE #
+    args.wakes = ["RHB_linear"]
+    args.n_t = 1
+    #############################################
+
+    # create model book
+    mbook = foxes.ModelBook()
+    ttype = foxes.models.turbine_types.PCtFile(args.turbine_file)
+    mbook.turbine_types[ttype.name] = ttype
+    mbook.turbine_models["kTI_K1"] = foxes.models.turbine_models.kTI(kTI=0.2, k_var="K1", ti_var=FV.AMB_TI)
+    mbook.turbine_models["kTI_K2"] = foxes.models.turbine_models.kTI(kTI=0.4, k_var="K2", ti_var=FV.AMB_TI)
+    mbook.wake_models["B_K1"] = foxes.models.wake_models.wind.BastankhahWake(
+        superposition="quadratic", k_var="K1"
+    )
+    mbook.wake_models["CH_K2"] = foxes.models.wake_models.ti.CrespoHernandezTIWake(
+        superposition="max", k_var="K2", use_ambti=False
+    )
+
+    # create states
+    states = foxes.input.states.SingleStateStates(
+        ws=args.ws, wd=args.wd, ti=args.ti, rho=args.rho
+    )
+
+    # create wind farm
+    print("\nCreating wind farm")
+    farm = foxes.WindFarm()
+    if args.layout is None:
+        foxes.input.farm_layout.add_row(
[…]
