+++ docs/source/output/FarmLayoutOutput.write_plot.rst
@@ -1,0 +1,10 @@
+..
+  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
+
+
+:hide-edit-link:
+
+
+
+.. python-apigen-entity-page::  foxes.output.FarmLayoutOutput.write_plot
+
