+++ docs/source/output/FarmResultsEval.calc_farm_yield.rst
@@ -1,0 +1,10 @@
+..
+  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
+
+
+:hide-edit-link:
+
+
+
+.. python-apigen-entity-page::  foxes.output.FarmResultsEval.calc_farm_yield
+

[…]
