+++ docs/source/conf.py
@@ -214,10 +214,10 @@ html_theme_options = {
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["_static"]
+#html_static_path = ["_static"]
 
 # custom.css is inside one of the html_static_path folders (e.g. _static)
-html_css_files = ["custom.css"]
+#html_css_files = ["custom.css"]
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
@@ -322,43 +322,43 @@ napoleon_attr_annotations = True
 # -- Extension configuration -------------------------------------------------
 
 python_apigen_modules = {
-    "foxes.core": "core/",
-    "foxes.data": "data/",
-    "foxes.input.farm_layout": "input/farm_layout/",
-    "foxes.input.states": "input/states/",
-    "foxes.input.windio": "input/windio/",
-    "foxes.output": "output/",
-    "foxes.models.model_book": "models/model_book/",
-    "foxes.models.farm_controllers": "models/farm_controllers/",
-    "foxes.models.farm_models": "models/farm_models/",
-    "foxes.models.partial_wakes": "models/partial_wakes/",
-    "foxes.models.point_models": "models/point_models/",
-    "foxes.models.rotor_models": "models/rotor_models/",
-    "foxes.models.turbine_models": "models/turbine_models/",
-    "foxes.models.turbine_types": "models/turbine_types/",
-    "foxes.models.vertical_profiles": "models/vertical_profiles/",
-    "foxes.models.wake_frames": "models/wake_frames/",
-    "foxes.models.wake_models": "models/wake_models/",
-    "foxes.models.wake_models.wind": "models/wake_models/wind/",
-    "foxes.models.wake_models.ti": "models/wake_models/ti/",
-    "foxes.models.wake_superpositions": "models/wake_superpositions/",
-    "foxes.utils": "utils/",
-    "foxes.utils.abl": "utils/abl",
-    "foxes.utils.geom2d": "utils/geom2d/",
-    "foxes.utils.runners": "utils/runners/",
-    "foxes.utils.two_circles": "utils/two_circles/",
-    "foxes.utils.abl.neutral": "utils/abl/neutral/",
-    "foxes.utils.abl.stable": "utils/abl/stable/",
-    "foxes.utils.abl.unstable": "utils/abl/unstable/",
-    "foxes.utils.abl.sheared": "utils/abl/sheared/",
-    "foxes.opt.core": "utils/opt/core/",
-    "foxes.opt.problems": "utils/opt/problems/",
-    "foxes.opt.problems.layout": "utils/opt/problems/layout",
-    "foxes.opt.problems.layout.geom_layouts": "utils/opt/problems/layout/geom_layouts",
-    "foxes.opt.problems.layout.geom_layouts.objectives": "utils/opt/problems/layout/geom_layouts/objectives",
-    "foxes.opt.problems.layout.geom_layouts.constraints": "utils/opt/problems/layout/geom_layouts/constraints",
-    "foxes.opt.objectives": "utils/opt/objectives/",
-    "foxes.opt.constraints": "utils/opt/constraints/",
+    "foxes.core": "_core/",
+    "foxes.data": "_data/",
+    "foxes.input.farm_layout": "_input/farm_layout/",
+    "foxes.input.states": "_input/states/",
+    "foxes.input.windio": "_input/windio/",
+    "foxes.output": "_output/",
+    "foxes.models.model_book": "_models/model_book/",
[…]
