+++ docs/source/models/partial_wakes/PartialGrid.contribute_to_wake_deltas.rst
@@ -1,10 +1,0 @@
-..
-  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
-
-
-:hide-edit-link:
-
-
-
-.. python-apigen-entity-page::  foxes.models.partial_wakes.PartialGrid.contribute_to_wake_deltas
-
