+++ docs/source/utils/opt/core/FarmOptProblem.evaluate_individual.rst
@@ -1,10 +1,0 @@
-..
-  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
-
-
-:hide-edit-link:
-
-
-
-.. python-apigen-entity-page::  foxes.opt.core.FarmOptProblem.evaluate_individual
-
