+++ docs/source/utils/opt/constraints/AreaGeometryConstraint.__str__.rst
@@ -1,10 +1,0 @@
-..
-  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
-
-
-:hide-edit-link:
-
-
-
-.. python-apigen-entity-page::  foxes.opt.constraints.AreaGeometryConstraint.__str__
-
