+++ docs/source/models/turbine_types/PCtFile.output_farm_vars.rst
@@ -1,10 +1,0 @@
-..
-  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
-
-
-:hide-edit-link:
-
-
-
-.. python-apigen-entity-page::  foxes.models.turbine_types.PCtFile.output_farm_vars
-
