+++ docs/source/models/turbine_models/Thrust2Ct.__init__.rst
@@ -1,10 +1,0 @@
-..
-  DO NOT EDIT. GENERATED by sphinx_immaterial.apidoc.python.apigen.
-
-
-:hide-edit-link:
-
-
-
-.. python-apigen-entity-page::  foxes.models.turbine_models.Thrust2Ct.__init__
-
