### 2023-06-12 a421f46
bugfix set_farm_vars.py and clean
