### 2023-06-12 0132cf9
dev setFarmVars for pop opt
