+++ foxes/core/point_data_model.py
@@ -103,19 +103,25 @@ class PointDataModelList(PointDataModel):
     `calculate` functions are called together
     under one common call of xarray's `apply_ufunc`.
 
-    Parameters
-    ----------
-    models: list of foxes.core.PointDataModel
-        The model list
-
     Attributes
     ----------
     models: list of foxes.core.PointDataModel
         The model list
 
+    :group: core
+
     """
 
     def __init__(self, models=[]):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        models: list of foxes.core.PointDataModel
+            The model list
+
+        """
         super().__init__()
         self.models = models
 
