+++ docs/source/conf.py
@@ -303,23 +303,7 @@ epub_title = project
 # A list of files that should not be packed into the epub file.
 epub_exclude_files = ["search.html"]
 
-# Napoleon settings
-napoleon_google_docstring = True
-napoleon_numpy_docstring = True
-napoleon_include_init_with_doc = True
-napoleon_include_private_with_doc = False
-napoleon_include_special_with_doc = True
-napoleon_use_admonition_for_examples = False
-napoleon_use_admonition_for_notes = False
-napoleon_use_admonition_for_references = False
-napoleon_use_ivar = False
-napoleon_use_param = True
-napoleon_use_rtype = True
-napoleon_preprocess_types = True
-napoleon_type_aliases = None
-napoleon_attr_annotations = True
-
-# -- Extension configuration -------------------------------------------------
+# -- python_apigen configuration -------------------------------------------------
 
 python_apigen_modules = {
     "foxes.core": "_core/",
@@ -360,9 +344,3 @@ python_apigen_modules = {
     "foxes.opt.objectives": "_utils/opt/objectives/",
     "foxes.opt.constraints": "_utils/opt/constraints/",
 }
-
-#python_apigen_default_groups = [
-#    ("class:.*", "Classes"),
-#    (r".*\.__(init|new)__", "Constructors"),
-#    (r".*\.__(str|repr)__", "String representation"),
-#]
