### 2023-06-12 #5464 @isdanni
> @isdanni & @antviro, could you please let us know where you stand with your reviews? We are now 4 and a half weeks into the review period and need to finish it soon.

Thanks for the follow up! Author(@SchmJo) has addressed my previously raised issues(just one minor left with some paper typos). Package functionality has also been confirmed. I have no further questions at the moment.
