+++ foxes/constants.py
@@ -1,88 +1,127 @@
-"""
-List of foxes constants.
-
-Attributes
-----------
-FARM : str
-    Wind farm identifier
-STATE : str
-    States identifier
-TURBINE : str
-    Wind turbine identifier
-TNAME : str
-    Wind turbine name identifier
-POINT : str
-    Point identifier
-POINTS : str
-    Points identifier
-RPOINT : str
-    Rotor point identifier
-RPOINTS : str
-    Rotor points identifier
-RWEIGHTS : str
-    Rotor point weights identifier
-AMB_RPOINT_RESULTS : str
-    Identified for ambient rotor point results
-VARS : str
-    Variables identifier
-VALID : str
-    Validity identifier
-TMODELS : str
-    Turbine models identifier
-TMODELS_SELS : str
-    Selected turbine models identifier
-DTYPE : alias
-    Default data type for floats
-ITYPE : alias
-    Default data type for int
-KAPPA : float
-    The Van-Karman constant
-W : str
-    The unit watt
-kW : str
-    The unit kilo watt
-MW : str
-    The unit mega watt
-GW : str
-    The unit giga watt
-TW : str
-    The unit terra watt
-P_UNITS : dict
-    Power unit factors relative to watts,
-    key: unit str, value: factor
-POP : str
-    Population identifier
-    
-"""
-
 import numpy as np
 
 FARM = "farm"
+""" Wind farm identifier
+:group: foxes.constants
+"""
+
 STATE = "state"
+""" States identifier
+:group: foxes.constants
+"""
+
 TURBINE = "turbine"
+""" Wind turbine identifier
+:group: foxes.constants
+"""
+
 TNAME = "tname"
+""" Wind turbine name identifier
+:group: foxes.constants
+"""
+
 POINT = "point"
+""" Point identifier
+:group: foxes.constants
+"""
+
 POINTS = "points"
+""" Points identifier
+:group: foxes.constants
+"""
+
 RPOINT = "rotor_point"
+""" Rotor point identifier
+:group: foxes.constants
+"""
+
 RPOINTS = "rotor_points"
+""" Rotor points identifier
+:group: foxes.constants
+"""
+
 RWEIGHTS = "rotor_weights"
+""" Rotor point weights identifier
+:group: foxes.constants
+"""
+
 AMB_RPOINT_RESULTS = "amb_rpoint_res"
+""" Identified for ambient rotor point results
+:group: foxes.constants
+"""
+
 
 VARS = "vars"
+""" Variables identifier
+:group: foxes.constants
+"""
+
 VALID = "valid"
+""" Validity identifier
+:group: foxes.constants
+"""
+
 TMODELS = "tmodels"
+""" Turbine models identifier
+:group: foxes.constants
+"""
+
 TMODEL_SELS = "tmodel_sels"
+"""Selected turbine models identifier
+:group: foxes.constants
+"""
+
 
 DTYPE = np.float64
+""" Default data type for floats
+:group: foxes.constants
+"""
+
 ITYPE = np.int64
+""" Default data type for int
+:group: foxes.constants
+"""
+
 
 KAPPA = 0.41
+""" The Van-Karman constant
+:group: foxes.constants
+"""
+
 
 W = "W"
+""" The unit watt
+:group: foxes.constants
+"""
+
 kW = "kW"
+""" The unit kilo watt
+:group: foxes.constants
+"""
+
 MW = "MW"
+""" The unit mega watt
+:group: foxes.constants
+"""
+
 GW = "GW"
+""" The unit giga watt
+:group: foxes.constants
+"""
+
 TW = "TW"
+""" The unit terra watt
[…]
