+++ docs/source/api.rst
@@ -4,15 +4,17 @@ API
     .. table:: 
         :widths: auto
 
-        ===================  ================================================
-        Package              Description
-        ===================  ================================================
-        :ref:`foxes`         Top-level package, shortcut to selected classes.
-        :ref:`foxes.core`    Core functionality and abstract base classes.
-        :ref:`foxes.input`   Functions and classes for user input.
-        :ref:`foxes.models`  All concrete model classes.
-        :ref:`foxes.utils`   Utilities and helper functions.
-        :ref:`foxes.opt`     Classes for wind farm optimization.
-        :ref:`foxes.output`  Classes and functions for output creation.
-        ===================  ================================================
+        ======================  ================================================
+        Package                 Description
+        ======================  ================================================
+        :ref:`foxes`            Top-level package, shortcut to selected classes.
+        :ref:`foxes.constants`  List of constants and identifiers.
+        :ref:`foxes.core`       Core functionality and abstract base classes.
+        :ref:`foxes.input`      Functions and classes for user input.
+        :ref:`foxes.models`     All concrete model classes.
+        :ref:`foxes.opt`        Classes for wind farm optimization.
+        :ref:`foxes.output`     Classes and functions for output creation.
+        :ref:`foxes.utils`      Utilities and helper functions.
+        :ref:`foxes.variables`  List of variables.
+        ======================  ================================================
 
