+++ docs/source/api_foxes.rst
@@ -1,7 +1,7 @@
 foxes
 =====
-The top-level package provides direct access to some 
-of selected classes and functions.
+The top-level package provides direct access to 
+selected classes and functions.
 
 All objects listed here can be reached directly
 via the *foxes* package, e.g. *foxes.WindFarm*.
