+++ examples/foxes/power_mask/run.py
@@ -8,6 +8,7 @@ import foxes.variables as FV
 import foxes.constants as FC
 from foxes.utils.runners import DaskRunner
 
+
 def run_foxes(args):
     cks = None if args.nodask else {FC.STATE: args.chunksize}
 
