+++ foxes/opt/problems/__init__.py
@@ -4,4 +4,3 @@ Wind farm optimization problems.
 from .opt_farm_vars import OptFarmVars
 
 from . import layout
-
