### 2023-06-13 78e9bd9
paper update, cf #38
