+++ setup.cfg
@@ -19,7 +19,7 @@ classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
-    Development Status :: 3 - Alpha
+    Development Status :: 4 - Beta
 
 [options]
 zip_safe = True
