### 2023-06-14 4038d2c
methods required for base class WakeModel added
