+++ foxes/models/wake_frames/__init__.py
@@ -3,5 +3,6 @@ Wake frame models.
 """
 from .rotor_wd import RotorWD
 from .streamlines import Streamlines
+from .timelines import Timelines
 from .yawed_wakes import YawedWakes
 from .farm_order import FarmOrder
