+++ foxes/models/wake_frames/timelines.py
@@ -1,0 +1,327 @@
+import numpy as np
+from scipy.interpolate import interpn
+from tqdm import tqdm
+
+from foxes.core import WakeFrame
+from foxes.utils import wd2uv
+from foxes.core.data import Data
+import foxes.variables as FV
+import foxes.constants as FC
+
+
+class Timelines(WakeFrame):
+    """
+    Streamline following wakes for timeseries based data
+
+    Attributes
+    ----------
+    cl_ipars: dict
+        Interpolation parameters for centre line
+        point interpolation
+
+    :group: models.wake_frames
+
+    """
+
+    def __init__(self, cl_ipars={}):
+        """
+        Constructor.
+
+        Parameters
+        ----------
+        cl_ipars: dict
+            Interpolation parameters for centre line
+            point interpolation
+
+        """
+        super().__init__()
+        self.cl_ipars = cl_ipars
+
+    def initialize(self, algo, verbosity=0):
+        """
+        Initializes the model.
+
+        This includes loading all required data from files. The model
+        should return all array type data as part of the idata return
+        dictionary (and not store it under self, for memory reasons). This
+        data will then be chunked and provided as part of the mdata object
+        during calculations.
+
+        Parameters
+        ----------
+        algo: foxes.core.Algorithm
+            The calculation algorithm
+        verbosity: int
+            The verbosity level, 0 = silent
+
+        Returns
+        -------
+        idata: dict
+            The dict has exactly two entries: `data_vars`,
+            a dict with entries `name_str -> (dim_tuple, data_ndarray)`;
+            and `coords`, a dict with entries `dim_name_str -> dim_array`
+
+        """              
+        idata = super().initialize(algo, verbosity)
+        
+        if verbosity > 0:
+            print(f"{self.name}: Pre-calculating streamlines")
+            pbar = tqdm(total=algo.n_states)
+        
+        # get and check times:
+        times = np.asarray(algo.states.index())
+        if not np.issubdtype(times.dtype, np.datetime64):
+            raise TypeError(f"{self.name}: Expecting state index of type np.datetime64, found {times.dtype}")
+        
+        # set initial coordinates to rotor centres:
+        data = np.zeros((algo.n_states, algo.n_turbines, 3), dtype=FC.DTYPE)
+        for ti, t in enumerate(algo.farm.turbines):
+            data[:, ti, :2] = t.xy
+            
+        # prepare mdata:
+        mdata = algo.idata_mem[algo.states.name]["data_vars"]
+        mdict = {v: d[1] for v, d in mdata.items()}
+        mdims = {v: d[0] for v, d in mdata.items()}
+        mdata = Data(mdict, mdims, loop_dims=[FC.STATE])
+        del mdict, mdims
+        
+        # prepare fdata:
+        fdata = Data({}, {}, loop_dims=[FC.STATE])
+        
+        # prepare pdata:
+        pdata = {v: np.zeros((algo.n_states, algo.n_turbines), dtype=FC.DTYPE) 
+                 for v in algo.states.output_point_vars(algo)}
+        pdata[FC.POINTS] = data
+        pdims = {FC.POINTS: (FC.STATE, FC.POINT, FV.XYH)}
+        pdims.update({v: (FC.STATE, FC.POINT) for v in pdata.keys()})
+        pdata = Data(pdata, pdims, loop_dims=[FC.STATE, FC.POINT])
+        
+        # calculate streamlines:
+        for si in range(1, algo.n_states):
+            
+            # reduced data by selecting single state:
+            hmdata, hpdata = algo.states.reduce_states([si], [mdata, pdata])
[…]
