+++ setup.cfg
@@ -40,6 +40,7 @@ install_requires =
     kaleido
     iwopy>=0.1.4
     pymoo>=0.6
+    tqdm
 
 [options.extras_require]
 test =
