### 2023-06-22 #5464 @antviro
Hi,
Sorry for the delay. I have not yet finished, but almost. I will try to do it along the day or tomorrow.

I am finishing running the examples, and actually I have to say that I am really liking it and, even the graphical output, which we tend to neglect (at least myself) is also quite polished and nice. I feel it is going to be a quite interesting publication for JOS.

Regarding the examples, I give you some preliminar feedback:
In the "Wind rose data" one https://fraunhoferiwes.github.io/foxes.docs/notebooks/wind_rose.html I get this error in the block 4:
```
In [5]: algo = foxes.algorithms.Downwind(
   ...:     mbook,
   ...:     farm,
   ...:     states=states,
   ...:     rotor_model="centre",
   ...:     wake_models=["Bastankhah_linear", "CrespoHernandez_max"],
   ...:     wake_frame="rotor_wd",
   ...:     partial_wakes_model="auto",
   ...:     chunks={FC.STATE: 500, FC.POINT: 4000},
   ...:     verbosity=0,
   ...: )
   ...:
Cell In[5], line 9
      1 algo = foxes.algorithms.Downwind(
      2     mbook,
      3     farm,
      4     states=states,
      5     rotor_model="centre",
      6     wake_models=["Bastankhah_linear", "CrespoHernandez_max"],
      7     wake_frame="rotor_wd",
      8     partial_wakes_model="auto",
----> 9     chunks={FC.STATE: 500, FC.POINT: 4000},
     10     verbosity=0,
     11 )

```
AttributeError: module 'foxes.constants' has no attribute 'STATE'

foxes.constants has also no POINT attribute.

I am using python 3.9.2 (running inside ipython 8.13.2): Is it possible the example comes from a previous version of foxes.constants?

Anyway, I have run without defining chunks, getting all the results. So the example is actually working (probably not parallelized if understood well the last comment of the example), but I just comment it to avoid an error appearing when someone runs it.

In the example Wind Sector Management, in the block 5

```
farm = foxes.WindFarm()
foxes.input.farm_layout.add_row(
    farm=farm,
    xy_base=[0.0, 0.0],
    xy_step=[600.0, 0.0],
    n_turbines=2,
    turbine_models=[ttype.name, "sector_rules", "PMask"],
)

ax = foxes.output.FarmLayoutOutput(farm).get_figure(figsize=(4, 3))
fig.show()
```
fig.show() is re-plotting last figure, where fig was defined, and not this one. Replacing fig by plt fixes it and leads to plot the layout.

Block 6 the chunks part produces the same fail as previous example (I solved commenting the line out and going on without chunks).

Best regards,
Antonio
