### 2023-06-22 #5464 @antviro
I have just realised I have all such variables in foxes.variables. So, probably I am running a newer version...
