### 2023-06-22 #5464 @SchmJo
BTW, you can always find the notebooks that are relevant for your version under ``foxes/notebooks`
