### 2023-06-22 #5464 @antviro
I see by default pip points out to 0.3.4. I run all examples just replacing constant with variables when it failed. Although now I have already installed 0.4.0 and have also checked there...

In my opinion is ready to be published as it is.
I think that perhaps including, not in the paper but in the API or documentation, some references about the  models included in book model: https://fraunhoferiwes.github.io/foxes.docs/api_models.html#foxes-models could improve its use.

Best regards and thank you for your efforts developing and releasing this code,
Antonio
