+++ foxes/models/model_book.py
@@ -169,6 +169,11 @@ class ModelBook:
                 base_frame=fm.wake_frames.Streamlines(step=s)
             )
         self.wake_frames["timelines"] = fm.wake_frames.Timelines()
+        self.wake_frames["timelines_1s"] = fm.wake_frames.Timelines(dt_min=1/60)
+        self.wake_frames["timelines_1min"] = fm.wake_frames.Timelines(dt_min=1)
+        self.wake_frames["timelines_10min"] = fm.wake_frames.Timelines(dt_min=10)
+        self.wake_frames["timelines_30min"] = fm.wake_frames.Timelines(dt_min=30)
+        self.wake_frames["timelines_60min"] = fm.wake_frames.Timelines(dt_min=60)
 
         self.wake_superpositions = Dict(
             name="wake_superpositions",
