+++ foxes/models/wake_frames/timelines.py
@@ -21,12 +21,15 @@ class Timelines(WakeFrame):
     cl_ipars: dict
         Interpolation parameters for centre line
         point interpolation
+    dt_min: float, optional
+        The delta t value in minutes, 
+        if not from timeseries data
 
     :group: models.wake_frames
 
     """
 
-    def __init__(self, max_wake_length=2e4, cl_ipars={}):
+    def __init__(self, max_wake_length=2e4, cl_ipars={}, dt_min=None):
         """
         Constructor.
 
@@ -37,11 +40,15 @@ class Timelines(WakeFrame):
         cl_ipars: dict
             Interpolation parameters for centre line
             point interpolation
+        dt_min: float, optional
+            The delta t value in minutes, 
+            if not from timeseries data
 
         """
         super().__init__()
         self.max_wake_length = max_wake_length
         self.cl_ipars = cl_ipars
+        self.dt_min = dt_min
 
     def initialize(self, algo, verbosity=0):
         """
@@ -75,8 +82,15 @@ class Timelines(WakeFrame):
         
         # get and check times:
         times = np.asarray(algo.states.index())
-        if not np.issubdtype(times.dtype, np.datetime64):
-            raise TypeError(f"{self.name}: Expecting state index of type np.datetime64, found {times.dtype}")
+        if self.dt_min is None:
+            if not np.issubdtype(times.dtype, np.datetime64):
+                raise TypeError(f"{self.name}: Expecting state index of type np.datetime64, found {times.dtype}")
+            elif len(times) == 1:
+                raise KeyError(f"{self.name}: Expecting 'dt_min' for single step timeseries")
+            dt = ( times[1:] - times[:-1] ).astype('timedelta64[s]').astype(FC.ITYPE)
+        else:
+            n = max(len(times)-1, 1)
+            dt = np.full(n, self.dt_min*60, dtype='timedelta64[s]').astype(FC.ITYPE)
 
         # calculate horizontal wind vector in all states:
         self._uv = np.zeros((algo.n_states, 1, 3), dtype=FC.DTYPE)
@@ -98,12 +112,14 @@ class Timelines(WakeFrame):
         pdims = {FC.POINTS: (FC.STATE, FC.POINT, FV.XYH)}
         pdims.update({v: (FC.STATE, FC.POINT) for v in pdata.keys()})
         pdata = Data(pdata, pdims, loop_dims=[FC.STATE, FC.POINT])
-        
+
         # calculate:      
         res = algo.states.calculate(algo, mdata, fdata, pdata)
-        dt = ( times[1:] - times[:-1] ).astype('timedelta64[s]').astype(FC.ITYPE)
-        self._dxy = wd2uv(res[FV.WD], res[FV.WS])[:-1, 0, :2] * dt[:, None]
-        self._dxy = np.insert(self._dxy, 0, self._dxy[0], axis=0)
+        if len(dt) == 1:
+            self._dxy = wd2uv(res[FV.WD], res[FV.WS])[:, 0, :2] * dt[:, None]
+        else:
+            self._dxy = wd2uv(res[FV.WD], res[FV.WS])[:-1, 0, :2] * dt[:, None]
+            self._dxy = np.insert(self._dxy, 0, self._dxy[0], axis=0)
 
         """ DEBUG
         import matplotlib.pyplot as plt
@@ -115,81 +131,6 @@ class Timelines(WakeFrame):
         """
         
         return idata
-
-    def _calc_coos(self, algo, mdata, fdata, points, tcase=False):
-        """
-        Helper function, calculates streamline coordinates
-        for given points.
-        """
-
-        # prepare:
-        n_states = mdata.n_states
[…]
