### 2023-06-27 #5464 @SchmJo
I can offer to drop the reference in "summary". It looks unusual to me, but I am not sure how you think about this. I leave this to you - I am happy with both options
