+++ foxes/output/flow_plots_2d.py
@@ -65,6 +65,7 @@ class FlowPlots2D(Output):
         ret_im,
         quiv=None,
         invert_axis=None,
+        animated=False,
     ):
         """
         Helper function for image creation
@@ -87,19 +88,20 @@ class FlowPlots2D(Output):
         # raw data image:
         if levels is None:
             im = hax.pcolormesh(
-                x_pos, y_pos, zz, vmin=vmin, vmax=vmax, shading="auto", cmap=cmap
+                x_pos, y_pos, zz, vmin=vmin, vmax=vmax, shading="auto", cmap=cmap, animated=animated,
             )
 
         # contour plot:
         else:
-            im = hax.contourf(x_pos, y_pos, zz, levels, vmax=vmax, vmin=vmin, cmap=cmap)
+            im = hax.contourf(x_pos, y_pos, zz, levels, vmax=vmax, vmin=vmin, cmap=cmap, animated=animated,)
 
+        qv = None
         if quiv is not None and quiv[0] is not None:
             n, pars, wd, ws = quiv
             uv = wd2uv(wd[si], ws[si])
             u = uv[:, 0].reshape([N_x, N_y]).T[::n, ::n]
             v = uv[:, 1].reshape([N_x, N_y]).T[::n, ::n]
-            hax.quiver(x_pos[::n], y_pos[::n], u, v, **pars)
+            qv = hax.quiver(x_pos[::n], y_pos[::n], u, v, animated=animated, **pars)
             del n, pars, u, v, uv
 
         hax.autoscale_view()
@@ -127,7 +129,7 @@ class FlowPlots2D(Output):
         if ret_state:
             out.append(si)
         if ret_im:
-            out.append(im)
+            out.append((im, qv) if qv is not None else (im,))
         if ret_state or ret_im:
             out = tuple(out)
 
@@ -162,6 +164,7 @@ class FlowPlots2D(Output):
         verbosity=0,
         ret_state=False,
         ret_im=False,
+        animated=False,
         **kwargs,
     ):
         """
@@ -223,6 +226,8 @@ class FlowPlots2D(Output):
             Flag for state index return
         ret_im: bool, optional
             Flag for image return
+        animated: bool
+            Switch for usage for an animation
         kwargs: dict, optional
             Parameters forwarded to the algorithm's calc_points
             function.
@@ -233,8 +238,9 @@ class FlowPlots2D(Output):
             The figure object
         si: int, optional
             The state index
-        im: matplotlib.collections.QuadMesh or matplotlib.QuadContourSet, optional
-            The image object
+        im: tuple, optional
+            The image objects, atplotlib.collections.QuadMesh
+            or matplotlib.QuadContourSet
 
         """
 
@@ -337,6 +343,7 @@ class FlowPlots2D(Output):
             vlabel,
             ret_state,
             ret_im,
+            animated=animated
         )
 
         return out
@@ -372,6 +379,7 @@ class FlowPlots2D(Output):
         verbosity=0,
         ret_state=False,
         ret_im=False,
+        animated=False,
         **kwargs,
     ):
         """
@@ -437,6 +445,8 @@ class FlowPlots2D(Output):
             Flag for state index return
         ret_im: bool, optional
             Flag for image return
+        animated: bool
+            Switch for usage for an animation
         kwargs: dict, optional
             Parameters forwarded to the algorithm's calc_points
             function.
@@ -447,8 +457,9 @@ class FlowPlots2D(Output):
             The figure object
         si: int, optional
             The state index
-        im: matplotlib.collections.QuadMesh or matplotlib.QuadContourSet, optional
-            The image object
+        im: tuple, optional
[…]
