+++ examples/foxes/timelines/run.py
@@ -2,118 +2,12 @@ import time
 import argparse
 import numpy as np
 import matplotlib.pyplot as plt
+import matplotlib.animation as animation
 
 import foxes
 import foxes.variables as FV
 import foxes.constants as FC
 from foxes.utils.runners import DaskRunner
-
-
-def run_foxes(args):
-    cks = None if args.nodask else {FC.STATE: args.chunksize}
-
-    mbook = foxes.models.ModelBook()
-    ttype = foxes.models.turbine_types.PCtFile(args.turbine_file)
-    mbook.turbine_types[ttype.name] = ttype
-
-    states = foxes.input.states.Timeseries(
-        data_source=args.states,
-        output_vars=[FV.WS, FV.WD, FV.TI, FV.RHO],
-        var2col={FV.WS: "ws", FV.WD: "wd", FV.TI: "ti"},
-        fixed_vars={FV.RHO: 1.225, FV.TI: 0.07},
-    )
-
-    farm = foxes.WindFarm()
-    N = int(args.n_turbines**0.5)
-    foxes.input.farm_layout.add_grid(
-        farm,
-        xy_base=np.array([0.0, 0.0]),
-        step_vectors=np.array([[3000.0, 0], [0, 4000.0]]),
-        steps=(N, N),
-        turbine_models=args.tmodels + [ttype.name],
-    )
-
-
-    if args.show_layout:
-        ax = foxes.output.FarmLayoutOutput(farm).get_figure()
-        plt.show()
-        plt.close(ax.get_figure())
-
-    Algo = foxes.algorithms.Iterative if args.iterative else foxes.algorithms.Downwind
-    algo = Algo(
-        mbook,
-        farm,
-        states=states,
-        rotor_model=args.rotor,
-        wake_models=args.wakes,
-        wake_frame=args.frame,
-        partial_wakes_model=args.pwakes,
-        chunks=cks,
-    )
-
-    time0 = time.time()
-
-    farm_results = algo.calc_farm(vars_to_amb=[FV.REWS, FV.P])
-
-    time1 = time.time()
-    print("\nCalc time =", time1 - time0, "\n")
-
-    o = foxes.output.FarmResultsEval(farm_results)
-    o.add_capacity(algo)
-    o.add_capacity(algo, ambient=True)
-    o.add_efficiency()
-
-    print("\nFarm results:\n")
-    print(farm_results)
-
-    # state-turbine results
-    farm_df = farm_results.to_dataframe()
-    print("\nFarm results data:\n")
-    print(
-        farm_df[
-            [
-                FV.X,
-                FV.WD,
-                FV.AMB_REWS,
-                FV.REWS,
-                FV.AMB_TI,
-                FV.TI,
-                FV.AMB_P,
-                FV.P,
-                FV.EFF,
-            ]
-        ]
-    )
-    print()
-
-    # results by turbine
-    turbine_results = o.reduce_states(
-        {
-            FV.AMB_P: "mean",
-            FV.P: "mean",
-            FV.AMB_CAP: "mean",
-            FV.CAP: "mean",
-            FV.EFF: "mean",
-        }
-    )
-    turbine_results[FV.AMB_YLD] = o.calc_turbine_yield(
-        algo=algo, annual=True, ambient=True
-    )
-    turbine_results[FV.YLD] = o.calc_turbine_yield(algo=algo, annual=True)
-    print("\nResults by turbine:\n")
-    print(turbine_results)
-
-    # power results
-    P0 = o.calc_mean_farm_power(ambient=True)
-    P = o.calc_mean_farm_power()
-    print(f"\nFarm power        : {P/1000:.1f} MW")
-    print(f"Farm ambient power: {P0/1000:.1f} MW")
[…]
