### 2023-06-28 830801a
Minor typo edit in JOSS paper (#39)
