### 2023-06-28 a49dbbc
adding input var function to wake models
