+++ foxes/models/wake_superpositions/linear.py
@@ -21,12 +21,14 @@ class LinearSuperposition(WakeSuperposition):
     lim_high: dict
         Higher limits of the final wake deltas. Key: variable str,
         value: float
+    svars: list of str
+        The scaling vafriables
 
     :group: models.wake_superpositions
 
     """
 
-    def __init__(self, scalings, lim_low=None, lim_high=None):
+    def __init__(self, scalings, lim_low=None, lim_high=None, svars=None):
         """
         Constructor.
 
@@ -44,6 +46,8 @@ class LinearSuperposition(WakeSuperposition):
         lim_high: dict, optional
             Higher limits of the final wake deltas. Key: variable str,
             value: float
+        svars: list of str, optional
+            The scaling vafriables
 
         """
         super().__init__()
@@ -51,6 +55,36 @@ class LinearSuperposition(WakeSuperposition):
         self.scalings = scalings
         self.lim_low = lim_low
         self.lim_high = lim_high
+        self.svars = svars
+
+    def input_farm_vars(self, algo):
+        """
+        The variables which are needed for running
+        the model.
+
+        Parameters
+        ----------
+        algo: foxes.core.Algorithm
+            The calculation algorithm
+
+        Returns
+        -------
+        input_vars: list of str
+            The input variable names
+
+        """
+        if self.svars is not None:
+            return self.svars
+        elif isinstance(self.scalings, dict):
+            return list(self.scalings.keys())
+        elif (
+            isinstance(self.scalings, str)
+            and len(self.scalings) > 15
+            and self.scalings[:15] == "source_turbine_"
+        ):
+            return [self.scalings[15:]]
+        else:
+            raise ValueError(f"{self.name}: Unable to determine scaling variable for scaling = '{self.scalings}'")
 
     def calc_wakes_plus_wake(
         self,
