### 2023-06-28 161b6b7
adding input var function to wake models
