+++ foxes/algorithms/iterative2/__init__.py
@@ -1,0 +1,3 @@
+from .iterative import Iterative2
+
+#from . import models
