+++ foxes/algorithms/__init__.py
@@ -3,6 +3,7 @@ Algorithms define the main calculation routines.
 """
 from .downwind.downwind import Downwind
 from .iterative.iterative import Iterative
+from .iterative2.iterative import Iterative2
 
 from . import downwind
 from . import iterative
