+++ foxes/algorithms/iterative2/iterative.py
@@ -2,8 +2,8 @@ from foxes.algorithms.downwind.downwind import Downwind
 
 class Iterative2(Downwind):
 
-    def __init__(self, max_it=100, **kwargs):
-        super().__init__(**kwargs)
+    def __init__(self, *args, max_it=100, **kwargs):
+        super().__init__(*args, **kwargs)
         self.max_it = max_it
 
     def calc_farm(
@@ -23,7 +23,8 @@ class Iterative2(Downwind):
 
             it += 1
 
-            break
+            if it > 1:
+                break
 
         return fres
 
